Mobile Widgets: Initial iOS Application
This is a native iOS application that accepts a single input (asset
ID/addr/handle) and renders back the NFT(s) from that blockchain
location.  Under the hood we are simply:

1. Allowing the user to input the locator
2. Verifying that the location (or the owner of the asset at the
   location) has a minimum # of NFTs for a required policy
3. Saving the location to a shared configuration value

The widget, which is a separate iOS target, is set to refresh every 15
minutes and do one of the following:
1. If an asset was specified, retrieve only the asset (refresh it)
2. If an address was specified, select a new random NFT to display

We support all widget size types and include default branding for
WildTangz so that this can be deployed to our hodlers.

[ Documentation: None ]
[ Testing: Built and deployed in Simulator and on phone ]