iOS Widget: Support for on-chain NFTs
The Swift module for Pool.PM needs to be able to handle base64-encoded
strings in case someone has rendered their data on chain.  So, to do
this we refactor the Pool.pm code to sniff out a base64 type (array of
strings), and parse it into a Data object.  The resultant Data object is
then rendered as with any PNG rendering from the Internet.

Note, we follow the boy scout rule here and make an update to the random
picker that has it try (up to 5 times) to pick a valid asset from the
user's wallet.  This will reduce the number of times overall that the
user will see the default Wild Tangz logo.

[ Documentation: None ]
[ Testing: Simulator, deploy to phone ]