Android App: Initial Android Native App Implementation
The vast majority of this commit is just the standard application
scaffolding produced by Android Studio.  Note that Android Studio is
built on IntelliJ (a fan favorite), so that is why you are seeing the
.idea/ files.

The key class in here is MainActivity.kt, the main application Activity
for the Android app, written in Kotlin.

The UI is created using @Composable tags (Jetpack Compose, an emerging
Android standard), and requires a Gradle upgrade to Android Material3
library version 1.0.0-alpha14.

Also note, we save the selected address to a simple shared configuration
preference rather than any long-term persistence for simplicity sake.

[ Documentation: None yet ]
[ Testing: Build to simulator ]