Android Widget: Create an authorization gateway
When the user clicks "enable", we want to ensure they can only load the
app and the corresponding widget if the wallet contains at least 1
Tangz.  So, to check this we launch a background task to retrieve
information from Pool.pm to verify the ownership.

NOTE: This requires some munging of the `MutableState` class in function
calls so that we can properly enable/disable the form while we are
checking this.

NOTE: We cannot make a custom Text Toast as per
https://itnext.io/android-11-toast-updates-7f1cd2245bc4 so we just punt
and make the simple text toast.

[ Documentation: None ]
[ Testing: Simulator ]