Android Widget: More intelligent memory handling in image decoding
Lots of NFTs are far too large for the Android WidgetView to display
(see
https://stackoverflow.com/questions/13494898/remoteviews-for-widget-update-exceeds-max-bitmap-memory-usage-error).
However, we cannot actually get the screen size from an AppWidget update
since it is a Context with no Display attached.  So, we simply cap the
photo size at 18MB in order to provide a reasonable limit for most
devices and only rescale when necessary.

[ Documentation: None ]
[ Testing: Simulator with large images ]