iOS Widget: Move to explicit photo memory management
Before we were relying on an "autorelease" pool to clean up old images
and avoid app crashes.  The only problem is that this was causing
several images that were too large to not be rendered at all.  Instead,
we use the ImageIO toolkit (WWDC18) to downscale the image if it is a
URL link to ensure that it renders correctly in the widget.

Code is adapted from here: https://developer.apple.com/forums/thread/109445

Note that there is significant refactoring to "do the right thing" based
on mediaType explicitly rather than do a fallback-load method which was
causing unexpected app behavior.

[ Documentation: None ]
[ Testing: Downscaled to very small size to check effect, TestFlight ]