Mobile Widgets: Support for Android 9+ and older iPhones
We want to support more users as we can, so we update the code to
compile against current API versions but be able to deployed to older
OSes.  Note that this requires us to lower the memory consumption for
Android and iOS modules, which we find acceptable so long as the image
resolution is high-def for a phone.

NOTE: We add an ITS exemption for the encryption we use (just standard
HTTPS) as per the instructions here:
https://developer.apple.com/documentation/security/complying_with_encryption_export_regulations#3145067

[ Documentation: None ]
[ Testing: Google internal test list and TestFlight to iPhone ]