Android v1.5 and iOS v1.3: Layout updates and code cleanups
The main changes made in here are to ensure that we have a set of
instructions for users.  Several users have complained that it is not
easy to understand widgets.  So, this app includes clearer directions
right from the main page of the app.

In addition, we reduce the requirements for the XCode build to iOS 14.0
to reach a broader swath of devices (though no lower due to the need for
specific SwiftUI APIs.

[ Documentation: None, instructions in app ]
[ Testing: Ran on simulators, TestFlight, and Google Play beta program ]