iOS Widget: Apewatch App Integration (v1.4)
This app upgrade creates a new widget in the "ApewatchProvider" file to:
1. Read in the user's data from apewatch.app
2. Compute summary statistics (portfolio value, # assets, # projects)
3. Render back to the user as a small or medium-sized widget

One added feature included in this commit is that if the user clicks on
the widget (small or medium), we deeplink them through the app directly
to the apewatch.app/ website in Safari to give them a visual to their
portfolio.

NOTE: This requires a synchronous implementation of `getAsJson` because
the calls to `Data(from: URL(urlString))` don't support any headers
which apewatch.app requires.

NOTE: This is gatekept on 4 Wild Tangz NFTs for the first iteration

[ Documentation: Updated instructions ]
[ Testing: simulator, TestFlight, external testing ]