iOS Widget: Support single native token selection and pagination (v1.4 build 4)
Commit d7e079a ignored the potential for pagination in the Blockfrost
results which can lead to users who have multiple assets in many parts
of their wallets being unable to either a) see those randomly or b)
authorize (if the WT is on a separate page).  This change rectifies that
by supporting pagination

In addition, we also add the ability for users to display single assets that are
native tokens even if there are multiples (only in the case they
specifically select this).

NOTE: During testing we discovered an error with metadata for Zeitaku
Dragons (e.g. asset1fvj5dtpzxnd5r88z9nvwsknjgxgl460gagg2r9) so we follow
the boy scout rule and fix the image processing there.

[ Documentation: None (all internal) ]
[ Testing: Simulator and TestFlight ]