iOS: Refactor AppAuthorization and extend to the widget viewer itself
This adds a single API call to Blockfrost but is worth it to avoid
unnecessarily loading up images if the user sold their Tangz.  However,
to follow the boy scout rule, we refactor into a single AppAuthorization
class in order to reduce the amount of locations where we are checking
static constants related to number of Tangz required to be held.

NOTE: We also change the refresh interval to lessen the amount of API
calls we are using with Blockfrost and avoid getting overcharged.

[ Documentation: Change in UI for if user sells their Tangz ]
[ Testing: Simulator, TestFlight ]