Android: Add authorization for widget prior to loading another photo
This ensures that we don't return any images from a wallet if the user
has transferred their Tangz out of the wallet. Following the boy scout
rule, we refactor the code in order to create a WalletAuth class that
can perform the authorization checks without exposing static constants
all over the application code.

[ Documentation: None ]
[ Testing: Simulator and Google Play beta program ]