iOS and Android: Update ApeWatch Widgets to new API
ApeWatch has kindly provided a new "skinnier" API so that the page will
load faster and be accessible for large portfolios.  This new API is
accessed under "attribute-and-link-licensing" via:

   https://apewatch.app/api/v1/account/<ADDR_OR_HANDLE>

NOTE: We choose between the "default" and "trait_floor" valuation method
for now but that is subject to change in the future depending on if
there is one that we are more comfortable with.

[ Documentation: None, no functionality change ]
[ Testing: Both platforms' simulators ]