feat(Valuation): Add support for TapTools API
ApeWatch has been officially deprecated so we can no longer use it.
Now, we migrate our app over to the TapTools API so that our users can
see their portfolio valuations.  Note that the number of projects is now
irrelevant and broken, though we anticipate fixing that shortly.
https://openapi.taptools.io/
https://coinmarketcap.com/api/documentation/v1

NOTE: CoinMarketCap is used as a free way to get pricing conversions.

[ Testing: Build for iOS 14 and use simulator ]